Refactor Settings System
Removed the value variable. Now instead, get_value() and set_value() functions are used for getting and setting values from a Field.

FieldOption field nodes now save as Dictionaries instead of an OptionItem class to the cfg file.

Updated main scene for using the new text based theme file and also changed how the settings are accessed for the themes.